+++ src/face_tracking/face_tracking.h
@@ -54,8 +54,6 @@ void face_tracking_copy_points(onnxmediapipe::FaceLandmarksResults *facelandmark
 
 face_tracking_bounding_box face_tracking_get_bounding_box(onnxmediapipe::FaceLandmarksResults *facelandmark_results, const unsigned short int *indices, int nb_indices);
 
-static void face_tracking_update(face_tracking_state *s);
-
 void face_tracking_create(face_tracking_state *s);
 
 void face_tracking_tick(face_tracking_state *s, obs_source_t *target_source, float deltatime);