+++ src/util/time_util.hpp
@@ -1,6 +1,6 @@
 #include <chrono>
 
-unsigned long get_time_ms() {
+inline unsigned long get_time_ms() {
     auto now = std::chrono::system_clock::now();
     auto now_ms = std::chrono::time_point_cast<std::chrono::milliseconds>(now);