+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Improved face tracking performance
+- Avoid multiple loading of the face tracking models
 - Added eye points in the face tracking texture, changed 468 → 478 in built-in shaders 
 - Face Detection Reference: add "Show texture" option
 - Various internal improvements